+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # Version 0.1.0 (unreleased)
--
+- convert foam dictionary to neofoam dictionary #13  [#13](https://github.com/exasim-project/FoamAdapter/pull/13)